[plugin.video.mediacorp] 3.0.2 (#968)
[plugin.video.mediacorp] 3.0.2
