+++ plugin.video.mediacorp/changelog.txt
@@ -1,2 +1,3 @@
+V3.0.2 website changes
 V3.0.1 Isengard version
 V2.0.1 Initial v2 release
