+++ plugin.video.mediacorp/resources/language/English/strings.po
@@ -1,0 +1,35 @@
+# Kodi Media Center language file
+# Addon Name: Mediacorp Singapore
+# Addon id: plugin.video.mediacorp
+# Addon Provider: t1m
+msgid ""
+msgstr ""
+"Project-Id-Version: XBMC Addons\n"
+"Report-Msgid-Bugs-To: alanwww1@xbmc.org\n"
+"POT-Creation-Date: YEAR-MO-DA HO:MI+ZONE\n"
+"PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
+"Last-Translator: Kodi Translation Team\n"
+"Language-Team: Amharic (http://www.transifex.com/projects/p/xbmc-addons/language/am/)\n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: 8bit\n"
+"Language: am\n"
+"Plural-Forms: nplurals=2; plural=(n > 1);\n"
+
+
+msgctxt "#30000"
+msgid "Video Resolution"
+msgstr ""
+
+msgctxt "#30001"
+msgid "SD"
+msgstr ""
+
+msgctxt "#30002"
+msgid "HD"
+msgstr ""
+
+msgctxt "#30003"
+msgid "No Playable Video Found"
+msgstr ""
+
