--- plugin.video.mediacorp/resources/language/English/strings.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<strings>
-       <string id="30000">Video Resolution</string>
-       <string id="30001">SD</string>
-       <string id="30002">HD</string>
-
-       <string id="30010">Sub-Titles</string>
-       <string id="30011">Enable Sub-Titles</string>
-       <string id="30012">Meta Data Cache</string>
-       <string id="30013">Enable Metadata Cache</string>
-       <string id="30014">Reset Metadata Cache</string>
-
-
-       <string id="30020">View</string>
-       <string id="30021">Enable Views</string>
-       <string id="30022">Default View</string>
-       <string id="30023">Movies View</string>
-       <string id="30024">Shows View</string>
-       <string id="30025">Episodes View</string>
-
-       <string id="30040">Metadata</string>
-       <string id="30041">Initialize Metadata Next Start</string>
-       <string id="30042">Initializing ....</string>
-
-       <string id="30100">tubitv</string>
-       <string id="30101">Initializing ....</string>
-
-
-</strings>
-
